Fix --secure-port setting
The Kubernetes control plane already defines a --secure-port flag, so
`rdd svc serve` should not define one as well; they are just competing
with setting default value, depending on load order, so created a bug
that looked like a race condition.

Also the actual (fallback) port needs to be determined when the control
plane is started, not when the service is created.

Finally we should always use 127.0.0.1 as the bind address to make sure
it is covered by the cert. Otherwise, if the local machine address changes,
the new address will cause cert validation issues.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>