Add script to modify `go.mod` to sync Kubernetes versions
Kubernetes modules tend to claim to require `v0.0.0` dependencies, and then
rely on `replace` directives to point to sources in the same monorepo.
Since we are not part of the monorepo, this makes things awkward; write a
script to force syncing everything `k8s.io/` to match `k8s.io/kubernetes`
in our `go.mod` (excluding the `.../v2` things, since those are versioned
separately).

Signed-off-by: Mark Yen <mark.yen@suse.com>