Refactor the notary-events BATS tests
This is based on a older refactoring that got lost in the rebase shuffle.

The previous events tests are not actually testing the right things, as
you can't test that no events are generated by setting the value repeatedly.
The calls get eliminated by the apiserver before they even reach the admission
controller. The way to test this is by making some other change to the resource,
like adding a label or annotation. Those can then generate duplicate events,
that in turn can be deduplicated by Kubernetes again.

But if we never produce duplicate events, then we cannot prove that they would
be de-duplicated.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>