CI: Fix getting write token
The setup to fetch the token is set up slightly differently; instead of
using a GitHub application, the token is stored directly in vault.
We also switch to running the workflow on pull_request instead of
pull_request_target and just let the action fail if it's from a cross-fork
pull request, since the main goal of this workflow is dependabot pull
requests which come from the same repository.

Signed-off-by: Mark Yen <mark.yen@suse.com>