CI: Disable setup-go caching in places
The setup-go action only uses the contents of go.sum to calculate the cache
key; this means that different workflows that download different modules
will end up with cache key collisions.  This means that the fastest job to
complete will end up being the one populating the cache; this typically
means the jobs that end up downloading the smallest amount of modules will
win (frequently, this means the license tag checking job).  This is not
ideal, because it means the cache is not being utilized effectively as we
will download more kubernetes modules more often.

To resolve this, manually mark the jobs that do not pull in the majority of
modules with `cache: false` to reserve the cache for the heavy weights.

Signed-off-by: Mark Yen <mark.yen@suse.com>