Refactor DeleteOwnedResources check all resource types
Enumerating resource types and resources inside a namspace is just a
local sqlite database query, and iterating over the results should be
just as quick as filtering stuff "server-side", which is the same
process on the same computer.

Not requiring the caller to pass in a list of resource types is more
maintenance friendly and the additional time required should be small,
given that resources are not constantly being deleted.

Label selectors are used only for setting up reconcilers and general
watches, and deletion is only controlled by owner references now.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>